+++ controllers/comments.controller.js
@@ -49,8 +49,21 @@ const updateComment = catchAsync(async (req, res) => {
   });
 });
 
+const createComment = catchAsync(async (req, res) => {
+  const { user_id, news_id, body } = req.body;
+  const comment = await Comments.create({ news_id, user_id, body });
+
+  res.status(201).json({
+      status: "true",
+      message: "Comentario creado con exito",
+      data: comment,
+  });
+});
+
 module.exports = {
   updateComment,
   deleteComment,
-  getAllComments
+  getAllComments,
+  createComment
+
 };
